use long

+++ openrlhf/trainer/ppo_utils/replay_buffer.py
@@ -113,11 +113,11 @@ def remove_padding_in_sequences(items):
             item.attention_mask,
             item.action_mask,
         )
-        right_pad = (1 - act_mask.int()).sum()
+        right_pad = (1 - act_mask.long()).sum()
         right_pad = None if right_pad == 0 else -right_pad
 
         # left_pad for seq and att_mask
-        left_pad = att_mask.argmax()
+        left_pad = att_mask.long().argmax()
         (
             item.sequences,
             item.action_log_probs,
